Split ABI in different sources
Since it is not feasible to create ABI functions that would work
across all ABI's, the tidiest way would be to split those across
several sources.

If we add more ABI's this would be relatively easy and give a good
overview of changes between the different ABI's instead of just
deduplicate everything.